+++ x64_dbg_dbg/variable.cpp
@@ -202,7 +202,7 @@ bool varset(const char* name, uint value, bool setreadonly)
 bool varset(const char* name, const char* string, bool setreadonly)
 {
     VAR_VALUE varvalue;
-    int size=strlen(string);
+    int size=(int)strlen(string);
     varvalue.size=size;
     varvalue.type=VAR_STRING;
     varvalue.u.data=new std::vector<unsigned char>;
